Refactoring Polio Plugin Api
