+++ iaso/api/common.py
@@ -19,6 +19,7 @@ from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet as BaseModelViewSet
 
 from hat.vector_control.models import APIImport
+from iaso.models import OrgUnit, OrgUnitType
 
 logger = logging.getLogger(__name__)
 
@@ -384,3 +385,23 @@ class CSVExportMixin:
             content=renderer.render(data),
         )
         return response
+
+
+class CustomFilterBackend(filters.BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        search = request.query_params.get("search")
+        if search:
+            country_types = OrgUnitType.objects.countries().only("id")
+            org_units = OrgUnit.objects.filter(
+                name__icontains=search, org_unit_type__in=country_types, path__isnull=False
+            ).only("id")
+
+            query = Q(obr_name__icontains=search) | Q(epid__icontains=search)
+            if len(org_units) > 0:
+                query.add(
+                    Q(initial_org_unit__path__descendants=OrgUnit.objects.query_for_related_org_units(org_units)), Q.OR
+                )
+
+            return queryset.filter(query)
+
+        return queryset
